requirements.txt
django=3.2.13
djangorestframework=3.13.1
pdfkit=1.0.0
qrcode=7.3.1